Refactored gates and binarytree to model
