+++ src/main/java/seedu/duke/model/gates/NotGate.java
@@ -1,4 +1,4 @@
-package seedu.duke.model.component.gates;
+package seedu.duke.model.gates;
 
 public class NotGate extends Gate {
