+++ src/main/java/seedu/duke/model/gates/OrGate.java
@@ -1,4 +1,4 @@
-package seedu.duke.model.component.gates;
+package seedu.duke.model.gates;
 
 public class OrGate extends TwoInputGate {
