+++ src/main/java/seedu/duke/model/gates/XorGate.java
@@ -1,4 +1,4 @@
-package seedu.duke.model.component.gates;
+package seedu.duke.model.gates;
 
 public class XorGate extends TwoInputGate {
     public XorGate(int input1, int input2) {
